refactor(tag-input): 优化目录结构
