+++ packages/devui-vue/devui/tag-input/src/icon-remove.tsx
